Order headers when generating CSV
It would be ideal if we could preserve the order of fields in the JSON
to the CSV, but there are a couple hurdles to this that make sorting the
CSV headers preferable:

1) Aeson (and every other JSON parser I've looked at) parses objects in
to an unordered map structure, so the field ordering is already lost
unless we implement our own JSON parser, which seems like overkill.

2) We'd still have to fix ordering for array values. Given an array of
objects, if the second has an array that's longer than the first then we
need to insert those headers in the right spot so that values under the
"same" array wind up in order.

Given this, it makes sense to sort the headers to ensure that the output
is consistent. This actually gives a major benefit as well: Semantically
identically JSON will produce identical CSV output.